requirements.txt
===================
[Packages List]
===================
Python     3.8.5
Flask      2.1.0
Werkzeug   2.1.2
joblib     1.0.0
pandas     1.4.2
numpy      1.21.6
matplotlib 3.3.4
sklearn    0.24.1
tensorflow 2.4.1
Pillow     8.4.0
ssl

=======================================
[Environment & Packages Installation]
=======================================

conda create -n cv_app_1 python=3.8.5
activate cv_app_1

pip install Flask==2.1.0
pip install Werkzeug==2.1.2
pip install joblib==1.0.0
pip install pandas==1.4.2
pip install numpy==1.21.6
pip install matplotlib==3.3.4
pip install scikit-learn==0.24.1
pip install tensorflow==2.4.1
pip install Pillow==8.4.0
pip install ssl